refactor: usando hook para obtenção do projeto passado o ID para que haja componentes mais stateless e limpos possíveis; adição do React.Fragment com o prop Key seguindo padrões e boas práticas do React
