+++ react-js/src/pages/Project.jsx
@@ -1,25 +1,11 @@
-import { useState, useEffect } from "react";
+import React from "react";
 import { useParams } from "react-router-dom";
-
-import projectData from "../assets/data/project/project.data";
-
 import Breadcrumb from "../components/Breadcrumb";
+import { useFetchProject } from "../hooks/useFetchProject";
 
 const Project = () => {
   const { id } = useParams();
-
-  const [project, setProject] = useState(null);
-  const [content, setContent] = useState('');
-
-  useEffect(() => {
-    const projectFound = projectData.find(project => project.id === Number(id));
-    setProject(projectFound);
-    getContent();
-
-    async function getContent() {
-      setContent(await projectFound.aboutContent);
-    }
-  }, [id]);
+  const { project, content } = useFetchProject(id);
 
   if (!project) return (<p>Carregando...</p>);
 
@@ -28,7 +14,7 @@ const Project = () => {
       <Breadcrumb previousPage="Portfolio" currentPage={project.title} />
       <article>
         <h1>{project.title}</h1>
-        <img class="foto-trabalhos" src={project.imagePreview.url} alt={project.imagePreview.title} title={project.imagePreview.title} />
+        <img className="foto-trabalhos" src={project.imagePreview.url} alt={project.imagePreview.title} title={project.imagePreview.title} />
 
         <h2>About Project</h2>
         <div dangerouslySetInnerHTML={{ __html: content }}></div>
@@ -37,9 +23,9 @@ const Project = () => {
         <p>
           {
             project.relatedLinks.map(item => (
-              <>
-                <a class="link" href={item.link} target="_blank">{item.title}</a><br />
-              </>
+              <React.Fragment key={item.title}>
+                <a className="link" href={item.link} target="_blank">{item.title}</a><br />
+              </React.Fragment>
             ))
           }
         </p>
